db 교체 및 Int 형변환

+++ testRecyclerView/app/src/main/java/org/techtown/testrecyclerview/search/SearchList.kt
@@ -51,7 +51,7 @@ class SearchList : AppCompatActivity() {
     private fun fillData() {
         var cursor: Cursor = db.rawQuery("SELECT * FROM real_nutri", null)
         while(cursor.moveToNext()) {
-            foodList.add(FoodData(cursor.getString(1), cursor.getString(2).toDouble(),100,cursor.getString(3).toDouble(), cursor.getString(4).toDouble(), cursor.getString(5).toDouble()))
+            foodList.add(FoodData(cursor.getString(1), cursor.getString(2).toInt(),100,cursor.getString(3).toInt(), cursor.getString(4).toInt(), cursor.getString(5).toInt()))
         }
 
 //        foodLiscursor.close()
